oauth2 authentication flow complete (now with multisite support)

+++ admin/class-collab-mendeley-plugin-admin.php
@@ -19,8 +19,8 @@
  * @package CollabMendeleyPluginAdmin
  * @author  Davide Parisi <davideparisi@gmail.com>
  */
-if ( ! class_exists( "Client" ) ) {
-	require_once plugin_dir_path( __FILE__ ) . "includes/vendor/autoload.php";
+if ( ! class_exists( "MendeleyApi" ) ) {
+	require_once plugin_dir_path( __DIR__ ) . "includes/class-mendeley-api-lib.php";
 }
 
 
@@ -74,10 +74,7 @@ class CollabMendeleyPluginAdmin {
 			return;
 		} */
 
-		$plugin             = CollabMendeleyPlugin::get_instance();
-		$this->plugin_slug  = $plugin->get_plugin_slug();
-		$this->options      = $this->get_options();
-		$this->callback_url = admin_url( 'options-general.php?page=' . $this->plugin_slug );
+		$this->init();
 
 
 		// Load admin style sheet and JavaScript.
@@ -98,10 +95,44 @@ class CollabMendeleyPluginAdmin {
 		 * http://codex.wordpress.org/Plugin_API#Hooks.2C_Actions_and_Filters
 		 */
 		add_action( 'admin_action_set_keys', array( $this, 'store_keys' ) );
+		add_action( 'admin_action_request_token', array( $this, 'request_access_token' ) );
 		add_filter( '@TODO', array( $this, 'filter_method_name' ) );
 
-		add_action( 'admin_init', array( $this, 'collab_mendeley_initialize_options' ) );
+		add_action( 'admin_init', array( $this, 'initialize_options' ) );
 
+	}
+
+	public function init(){
+		$plugin             = CollabMendeleyPlugin::get_instance();
+		$this->plugin_slug  = $plugin->get_plugin_slug();
+		$this->callback_url = admin_url( 'options-general.php?page=' . $this->plugin_slug );
+		$this->client       = new MendeleyApi();
+		$this->options      = $this->get_options();
+		if (isset($this->options['access_token'])){
+			$this->check_access_token();
+		}
+	}
+
+	public function check_access_token(){
+		$options = $this->options;
+		$result = $options['access_token']['result'];
+		$expire_time = (time() + $result['expire_in']);
+		if ( time() > $expire_time){
+			$this->refresh_token();
+		}
+	}
+
+	public function refresh_token(){
+		$options = $this->options;
+		$result = $options['access_token']['result'];
+		$refresh_token = $result['refresh_token'];
+		$this->client->set_client_id($options['client_id']);
+		$this->client->set_client_secret($options['client_secret']);
+		$this->client->set_callback_url($this->callback_url);
+		$this->client->init();
+		$new_token = $this->client->refresh_access_token($refresh_token);
+		$options['access_token'] = $new_token;
+		$this->update_options($options);
 	}
 
 	/**
@@ -170,23 +201,36 @@ class CollabMendeleyPluginAdmin {
 
 	}
 
-	public function collab_mendeley_initialize_options() {
+	/* ------------------------------------------------------------------------ *
+    * Setting Registration
+    * ------------------------------------------------------------------------ */
+
+	public function default_keys_options() {
+		$defaults = array(
+			'client_id'     => '',
+			'client_secret' => ''
+		);
+
+		return apply_filters( 'default_keys_options', $defaults );
+	}
+
+	public function initialize_options() {
 
 		// check if multisite environment
 		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
-			if (false === get_site_option($this->plugin_slug)){
-				add_site_option($this->plugin_slug, array('client_id' =>'', 'client_secret' => '' ));
+			if ( false === get_site_option( $this->plugin_slug ) ) {
+				add_site_option( $this->plugin_slug, apply_filters( 'default_keys_options', $this->default_keys_options() ) );
 			}
 		} else {
-			if (false === get_option($this->plugin_slug)){
-				add_option($this->plugin_slug, array('client_id' =>'', 'client_secret' => '' ));
+			if ( false === get_option( $this->plugin_slug ) ) {
+				add_option( $this->plugin_slug, apply_filters( 'default_keys_options', $this->default_keys_options() ) );
 			}
 		}
 
 		add_settings_section(
 			'collab_mendeley_settings_section',
-			'Collab Mendeley Setting',
-			array($this,'collab_mendeley_options_callback'),
+			'API Key Setting',
+			array( $this, 'options_callback' ),
 			$this->plugin_slug
 		);
 
@@ -208,32 +252,52 @@ class CollabMendeleyPluginAdmin {
 			array( 'Insert the client secret' )
 		);
 
-		register_setting( $this->plugin_slug, $this->plugin_slug );
+		register_setting(
+			$this->plugin_slug,
+			$this->plugin_slug,
+			array( $this, 'validate' )
+		);
 	}
 
-	public function collab_mendeley_options_callback() {
-		echo '<p class="description">Demo</p>';
+	public function options_callback() {
+		echo '<p class="description">Insert the <code>client ID</code> and <code>client secret</code> you have got from registering this plugin on <a href="http://dev.mendeley.com">Mendeley</a></p>';
 	}
 
 	public function client_id_input_callback( $args ) {
-		if(function_exists('is_multisite') && is_multisite()){
-			$options = get_site_option($this->plugin_slug);
-		}else{
+		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			$options = get_site_option( $this->plugin_slug );
+		} else {
 			$options = get_option( $this->plugin_slug );
 		}
-		$html    = '<input type="text" id="client_id" name="' . $this->plugin_slug . '[client_id]" value="' . $options['client_id'] . '"/>';
+		$html = '<input type="text" id="client_id" name="' . $this->plugin_slug . '[client_id]" value="' . $options['client_id'] . '" />'; // readonly="'. (isset($options['client_id']) ? "true" : "false")  .'"
 		echo $html;
 	}
 
 	public function client_secret_input_callback( $args ) {
-		if(function_exists('is_multisite') && is_multisite()){
-			$options = get_site_option($this->plugin_slug);
-		}else{
+		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			$options = get_site_option( $this->plugin_slug );
+		} else {
 			$options = get_option( $this->plugin_slug );
 		}
 		$options = get_option( $this->plugin_slug );
-		$html    = '<input type="text" id="client_secret" name="' . $this->plugin_slug . '[client_secret]" value="' . $options['client_secret'] . '"/>';
+		$html    = '<input type="text" id="client_secret" name="' . $this->plugin_slug . '[client_secret]" value="' . $options['client_secret'] . '" />'; // readonly="'. (isset($options['client_id']) ? "true" : "false") .'"
 		echo $html;
+	}
+
+	public function validate( $input ) {
+		$output = array();
+		foreach ( $input as $key => $value ) {
+			if ( isset( $input[ $key ] ) ) {
+				if ( $key == 'access_token' ) {
+					$output[ $key ] = $input[ $key ];
+				} else {
+					$output[ $key ] = strip_tags( stripslashes( $input[ $key ] ) );
+				}
+			}
+		}
+
+		return apply_filters( 'validate', $output, $input );
+
 	}
 
 	/**
@@ -266,6 +330,9 @@ class CollabMendeleyPluginAdmin {
 	 * @since    1.0.0
 	 */
 	public function display_plugin_admin_page() {
+		if ( isset( $_GET['code'] ) ) {
+			$this->store_access_token( $_GET['code'] );
+		}
 		include_once( 'views/admin.php' );
 	}
 
@@ -299,11 +366,45 @@ class CollabMendeleyPluginAdmin {
 		// @TODO: Define your filter hook callback here
 	}
 
+	public function request_access_token() {
+		$options = array();
+		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			$options = get_site_option( $this->plugin_slug );
+		} else {
+			$options = get_option( $this->plugin_slug );
+		}
+		if ( $options['client_id'] === '' || $options['client_secret'] === '' ) {
+			//@todo: do something if keys are void
+			exit();
+		}
+		if ( ! isset( $this->client ) ) {
+			$this->client = new MendeleyApi();
+		}
+		$this->client->set_client_id( $options['client_id'] );
+		$this->client->set_client_secret( $options['client_secret'] );
+		$this->client->set_callback_url( $this->callback_url );
+		$this->client->init();
+		$this->client->start_authorization_flow();
+	}
+
+	public function store_access_token( $auth_code ) {
+		$options = $this->get_options();
+		$this->client->set_client_id( $options['client_id'] );
+		$this->client->set_client_secret( $options['client_secret'] );
+		$this->client->set_callback_url( $this->callback_url );
+		$this->client->init();
+		$access_token            = $this->client->get_access_token( $auth_code );
+		$options['access_token'] = $access_token;
+		$this->update_options( $options );
+
+	}
+
 	/*------------------------------------------------------------------------------
 	 *
 	 * Private Functions
 	 *
 	 -----------------------------------------------------------------------------*/
+
 
 	/**
 	 * Update options array with db data (if present)
@@ -311,22 +412,14 @@ class CollabMendeleyPluginAdmin {
 	 * @return null
 	 */
 	private function get_options() {
-		// if $options is already present return $options
-		if ( isset( $this->options ) ) {
-			return $this->options;
-		}
-		// check if options are in the db and store them in $this->options
-		$tmp_options = get_option( $this->plugin_slug );
-		if ( isset( $tmp_options ) ) {
-			return $tmp_options;
+		$opts = array();
+		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			$opts = get_site_option( $this->plugin_slug );
 		} else {
-			// otherwise initialize to an empty array
-			$this->options = array();
-			add_option( $this->plugin_slug, $this->options );
-
-			return $this->options;
+			$opts = get_option( $this->plugin_slug );
 		}
 
+		return $opts;
 	}
 
 	/**
@@ -335,8 +428,11 @@ class CollabMendeleyPluginAdmin {
 	 * @param $options
 	 */
 	private function update_options( $options ) {
-		// #TODO: check if db options are stale and then update
-		update_option( $this->plugin_slug, $options );
+		if ( function_exists( 'is_multisite' ) && is_multisite() ) {
+			update_site_option( $this->plugin_slug, $options );
+		} else {
+			update_option( $this->plugin_slug, $options );
+		}
 	}
 
 
@@ -421,16 +517,22 @@ class CollabMendeleyPluginAdmin {
 	}
 
 	/**
-	 * NOTE:     Actions are points in the execution of a page or process
-	 *           lifecycle that WordPress fires.
+	 * Generic function to show a message to the user using WP's
+	 * standard CSS classes to make use of the already-defined
+	 * message colour scheme.
 	 *
-	 *           Actions:    http://codex.wordpress.org/Plugin_API#Actions
-	 *           Reference:  http://codex.wordpress.org/Plugin_API/Action_Reference
-	 *
-	 * @since    1.0.0
+	 * @param $message The message you want to tell the user.
+	 * @param $errormsg If true, the message is an error, so use
+	 * the red message style. If false, the message is a status
+	 * message, so use the yellow information message style.
 	 */
-	public function store_keys() {
+	function showMessage( $message, $errormsg = false ) {
+		if ( $errormsg ) {
+			echo '<div id="message" class="error">';
+		} else {
+			echo '<div id="message" class="updated fade">';
+		}
 
+		echo "<p><strong>$message</strong></p></div>";
 	}
-
 }
